+++ libs/netty4/build.gradle
@@ -28,128 +28,36 @@
  * under the License.
  */
 
-
-import org.opensearch.gradle.info.BuildParams
-import org.opensearch.gradle.test.RestIntegTestTask
-import org.opensearch.gradle.test.TestTask
-import org.opensearch.gradle.test.rest.JavaRestTestPlugin
-import org.opensearch.gradle.test.InternalClusterTestPlugin
-
-apply plugin: 'opensearch.yaml-rest-test'
-apply plugin: 'opensearch.java-rest-test'
-apply plugin: 'opensearch.internal-cluster-test'
-apply from: "$rootDir/gradle/fips.gradle"
-
-// The transport-netty4 plugin is published to maven
 apply plugin: 'opensearch.publish'
 
-/*
- TODOs:
-   * fix permissions such that only netty4 can open sockets etc?
-   * fix the hack in the build framework that copies transport-netty4 into the integ test cluster
-   * maybe figure out a way to run all tests from core with netty4/network?
- */
-opensearchplugin {
-  description = 'Netty 4 based transport implementation'
-  classname = 'org.opensearch.transport.Netty4ModulePlugin'
-  hasClientJar = true
-}
-
 dependencies {
-  api project(':libs:opensearch-netty4')
+  api project(':libs:opensearch-common')
+  api project(':libs:opensearch-core')
 
-  // network stack
-  api "io.netty:netty-buffer:${versions.netty}"
-  api "io.netty:netty-codec-base:${versions.netty}"
-  api "io.netty:netty-codec-compression:${versions.netty}"
-  api "io.netty:netty-codec:${versions.netty}"
-  api "io.netty:netty-codec-http:${versions.netty}"
-  api "io.netty:netty-codec-http2:${versions.netty}"
-  api "io.netty:netty-codec-http3:${versions.netty}"
-  api "io.netty:netty-codec-classes-quic:${versions.netty}"
-  api "io.netty:netty-codec-native-quic:${versions.netty}"
-  api "io.netty:netty-common:${versions.netty}"
-  api "io.netty:netty-handler:${versions.netty}"
-  api "io.netty:netty-resolver:${versions.netty}"
-  api "io.netty:netty-transport:${versions.netty}"
-  api "io.netty:netty-transport-native-unix-common:${versions.netty}"
+  api libs.netty.handler
+  api libs.netty.buffer
+  api libs.netty.common
+  api libs.netty.transport
+  api libs.netty.codec.classes.quic
 
-  testFipsRuntimeOnly "org.bouncycastle:bc-fips:${versions.bouncycastle_jce}"
-  testFipsRuntimeOnly "org.bouncycastle:bctls-fips:${versions.bouncycastle_tls}"
-  testFipsRuntimeOnly "org.bouncycastle:bcutil-fips:${versions.bouncycastle_util}"
+  api libs.lucene.core;
 
-  // Bundle all supported OSes and Archs
-  runtimeOnly "io.netty:netty-codec-native-quic:${versions.netty}:linux-x86_64"
-  runtimeOnly "io.netty:netty-codec-native-quic:${versions.netty}:linux-aarch_64"
-  runtimeOnly "io.netty:netty-codec-native-quic:${versions.netty}:osx-x86_64"
-  runtimeOnly "io.netty:netty-codec-native-quic:${versions.netty}:osx-aarch_64"
-  runtimeOnly "io.netty:netty-codec-native-quic:${versions.netty}:windows-x86_64"
-}
-
-restResources {
-  restApi {
-    includeCore '_common', 'cluster', 'nodes'
-  }
+  testImplementation project(":test:framework")
 }
 
 tasks.named("dependencyLicenses").configure {
   mapping from: /netty-.*/, to: 'netty'
+  mapping from: /lucene-.*/, to: 'lucene'
 }
 
-test {
-  /*
-   * We have to disable setting the number of available processors as tests in the same JVM randomize processors and will step on each
-   * other if we allow them to set the number of available processors as it's set-once in Netty.
-   */
-  systemProperty 'opensearch.set.netty.runtime.available.processors', 'false'
+tasks.named('forbiddenApisMain').configure {
+  // nio does not depend on core, so only jdk signatures should be checked
+  // es-all is not checked as we connect and accept sockets
+  replaceSignatureFiles 'jdk-signatures'
 }
-
-internalClusterTest {
-  systemProperty 'opensearch.set.netty.runtime.available.processors', 'false'
-}
-
-javaRestTest {
-  systemProperty 'opensearch.set.netty.runtime.available.processors', 'false'
-}
-
-TaskProvider<Test> pooledTest = tasks.register("pooledTest", TestTask) {
-  include '**/*Tests.class'
-  systemProperty 'opensearch.set.netty.runtime.available.processors', 'false'
-  systemProperty 'opensearch.use_unpooled_allocator', 'false'
-}
-
-TaskProvider<Test> pooledInternalClusterTest = tasks.register("pooledInternalClusterTest", TestTask) {
-  include '**/*IT.class'
-  systemProperty 'opensearch.set.netty.runtime.available.processors', 'false'
-  systemProperty 'opensearch.use_unpooled_allocator', 'false'
-  SourceSetContainer sourceSets = project.getExtensions().getByType(SourceSetContainer.class);
-  SourceSet internalTestSourceSet = sourceSets.getByName(InternalClusterTestPlugin.SOURCE_SET_NAME)
-  setTestClassesDirs(internalTestSourceSet.getOutput().getClassesDirs())
-  setClasspath(internalTestSourceSet.getRuntimeClasspath())
-}
-
-RestIntegTestTask pooledJavaRestTest = tasks.create("pooledJavaRestTest", RestIntegTestTask) {
-    systemProperty 'opensearch.set.netty.runtime.available.processors', 'false'
-    SourceSetContainer sourceSets = project.getExtensions().getByType(SourceSetContainer.class);
-    SourceSet javaRestTestSourceSet = sourceSets.getByName(JavaRestTestPlugin.SOURCE_SET_NAME)
-    setTestClassesDirs(javaRestTestSourceSet.getOutput().getClassesDirs())
-    setClasspath(javaRestTestSourceSet.getRuntimeClasspath())
-}
-testClusters.pooledJavaRestTest {
-  systemProperty 'opensearch.use_unpooled_allocator', 'false'
-}
-check.dependsOn(pooledTest, pooledJavaRestTest, pooledInternalClusterTest)
 
 thirdPartyAudit {
   ignoreMissingClasses(
-    'com.aayushatharva.brotli4j.Brotli4jLoader',
-    'com.aayushatharva.brotli4j.decoder.DecoderJNI$Status',
-    'com.aayushatharva.brotli4j.decoder.DecoderJNI$Wrapper',
-    'com.aayushatharva.brotli4j.encoder.BrotliEncoderChannel',
-    'com.aayushatharva.brotli4j.encoder.Encoder$Mode',
-    'com.aayushatharva.brotli4j.encoder.Encoder$Parameters',
-    // classes are missing
-
     // from io.netty.logging.CommonsLoggerFactory (netty)
     'org.apache.commons.logging.Log',
     'org.apache.commons.logging.LogFactory',
@@ -184,19 +92,15 @@ thirdPartyAudit {
     'org.slf4j.LoggerFactory',
     'org.slf4j.spi.LocationAwareLogger',
 
-    'com.ning.compress.BufferRecycler',
-    'com.ning.compress.lzf.ChunkDecoder',
-    'com.ning.compress.lzf.ChunkEncoder',
-    'com.ning.compress.lzf.LZFChunk',
-    'com.ning.compress.lzf.LZFEncoder',
-    'com.ning.compress.lzf.util.ChunkDecoderFactory',
-    'com.ning.compress.lzf.util.ChunkEncoderFactory',
-    'lzma.sdk.lzma.Encoder',
-    'net.jpountz.lz4.LZ4Compressor',
-    'net.jpountz.lz4.LZ4Factory',
-    'net.jpountz.lz4.LZ4FastDecompressor',
-    'net.jpountz.xxhash.XXHash32',
-    'net.jpountz.xxhash.XXHashFactory',
+    'io.netty.channel.unix.FileDescriptor',
+    'io.netty.channel.unix.UnixChannel',
+    'io.netty.handler.codec.ByteToMessageDecoder',
+    'io.netty.handler.codec.ByteToMessageDecoder$Cumulator',
+    'io.netty.handler.codec.DecoderException',
+    'io.netty.handler.codec.base64.Base64',
+    'io.netty.handler.codec.base64.Base64Dialect',
+    'io.netty.resolver.AddressResolver',
+    'io.netty.resolver.AddressResolverGroup',
     'io.netty.internal.tcnative.AsyncSSLPrivateKeyMethod',
     'io.netty.internal.tcnative.AsyncTask',
     'io.netty.internal.tcnative.CertificateCallback',
@@ -207,6 +111,19 @@ thirdPartyAudit {
     'io.netty.internal.tcnative.SSL',
     'io.netty.internal.tcnative.SSLSession',
     'io.netty.internal.tcnative.SSLSessionCache',
+    'io.netty.resolver.DefaultAddressResolverGroup',
+
+    'jdk.incubator.vector.ByteVector',
+    'jdk.incubator.vector.DoubleVector',
+    'jdk.incubator.vector.FloatVector',
+    'jdk.incubator.vector.IntVector',
+    'jdk.incubator.vector.LongVector',
+    'jdk.incubator.vector.ShortVector',
+    'jdk.incubator.vector.Vector',
+    'jdk.incubator.vector.VectorMask',
+    'jdk.incubator.vector.VectorOperators',
+    'jdk.incubator.vector.VectorShape',
+    'jdk.incubator.vector.VectorSpecies',
 
     'org.conscrypt.AllocatedBuffer',
     'org.conscrypt.BufferAllocator',
