+++ modules/transport-netty4/src/main/java/org/opensearch/http/netty4/Netty4Http3ServerTransport.java
@@ -12,6 +12,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.ExceptionsHelper;
 import org.opensearch.OpenSearchException;
+import org.opensearch.common.Randomness;
 import org.opensearch.common.network.NetworkService;
 import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.settings.Setting;
@@ -291,7 +292,7 @@ public class Netty4Http3ServerTransport extends AbstractHttpServerTransport {
                     .initialMaxStreamDataBidirectionalLocal(SETTING_H3_MAX_STREAM_LOCAL_LENGTH.get(settings).getBytes())
                     .initialMaxStreamDataBidirectionalRemote(SETTING_H3_MAX_STREAM_REMOTE_LENGTH.get(settings).getBytes())
                     .initialMaxStreamsBidirectional(SETTING_H3_MAX_STREAMS.get(settings).longValue())
-                    .tokenHandler(new SecureQuicTokenHandler())
+                    .tokenHandler(new SecureQuicTokenHandler(Randomness.createSecure()))
                     .handler(new ChannelInitializer<QuicChannel>() {
                         @Override
                         protected void initChannel(QuicChannel ch) {
